Fixed errors related to `arguments.callee.caller`. Fixed the capability lib, updated dependency, moved the caller check into the non-v8 adapter code, so it will gracefully degrade by strict mode only environments. This has a price, we cannot override the present stack capturing methods by UserError descendants, since it is important to keep the 4 frame distance constant between the `CustomError.prototype.{constructor|clone}` and the `FrameStringSource.prototype.createError`. For example overriding the `factory` and calling `UserError.constructor` with `UserError.call(this)` in order to create the `stack` would result in an additional frame which captures the `CustomError.constructor`. The same stands on the `prototype.clone` method, since these methods start the capturing process. This needs to be fixed, but I don't know yet how to do that. I added a warning about this to the documentation. That's all I can do so far, it's still better than it was before. #21
Checked `Object.defineProperty`, but I did not find any bugs. I added the `configurable: false`, so it will be evident now, that the `stack` property is non-writable and non-configurable, we can call it frozen. I froze the `Error.prepareStackTrace` method too, so using another lib, which overrides the same method will cause probably an Error if somebody wants to configure it with ˙Object.defineProperty˙. Trying to write it will ignore the command, so it won't be that obvious what's causing the problem, but that's not my business, it is documented. #15 #22